requirements.txt
pbr>=5.2.0

Babel>=2.6.0
decorator>=4.4.0
django-rest-swagger>=2.2.0
djangorestframework>=3.9.4
jsonfield>=2.0.2
keystoneauth1>=3.14.0
keystonemiddleware>=6.0.0
python-cinderclient>=4.1.0
python-keystoneclient>=3.19.0
python-neutronclient>=6.12.0
python-novaclient>=14.0.0
python-octaviaclient>=1.8.0
PyYAML>=5.1
six>=1.12.0
confspirator>=0.1.6

# Address soon:
Django>=1.11,<1.12
# was pinned because of django 1.11
mysqlclient>=1.3.10,<1.4
